apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mywebback-api"
version = "0.1.0"
description = "FastAPI backend for the monorepo starter"
readme = "README.md"
authors = [{ name = "Monorepo Starter", email = "dev@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.105.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic-settings>=2.0.3",
    "pydantic[email]>=2.6.1",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "psycopg2-binary>=2.9.9",
    "pymysql>=1.1.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.8",
    "pytest>=7.4.3"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["app/tests"]
